refactor(loader): strip findLoader
